+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - upgrade flutter sdk to 3.22.0
 - fix export of `ListView`
+- add support for widgets like `CustomPaint` and `Charts` by introducing a `CaptureWrapper` widget
 
 ## 0.2.0